backend/FoodVisorAPI/requirements.txt
awsebcli==3.20.10
blessed==1.20.0
blinker==1.8.2
botocore==1.31.85
cement==2.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.3
Flask==3.0.3
Flask-Cors==4.0.1
idna==3.7
importlib_metadata==8.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
pathspec==0.10.1
pillow==10.4.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.32.3
semantic-version==2.8.5
six==1.16.0
termcolor==1.1.0
urllib3==1.26.19
wcwidth==0.1.9
Werkzeug==3.0.3
zipp==3.19.2
